udf/anomaly-detection/pyproject.toml
[tool.poetry]
name = "anomaly-detection"
version = "0.3.2"
description = "Numalogic UDFs to do anomaly detection on timeseries data"
authors = ["Numalogic developers"]
packages = [{ include = "anomalydetection" }]
maintainers = [
    "Avik Basu <avikbasu93@gmail.com>",
    "Nandita Koppisetty <nandita.iitkgp@gmail.com>",
]
classifiers = [
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10"
]
repository = "https://github.com/numaproj/numalogic"

[tool.poetry.dependencies]
python = "~3.10"
redis = "^4.3.1"
pynumaflow = "~0.4.0"
aiorun = "^2022.11.1"
numalogic = {version = "~0.3.7", extras = ["mlflow"]}
boto3 = "^1.25.2"
orjson = "^3.8.4"
omegaconf = "^2.3.0"
watchdog = "^3.0.0"

[tool.poetry.group.mlflowserver]
optional = true

[tool.poetry.group.mlflowserver.dependencies]
mlflow = "^1.30.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytorch-lightning = "^1.8.6"
coverage = "^6.3"
black = "^22.1"
fakeredis = "^1.8"
flake8 = "^5.0"
pytest = "^7.1"
pytest-cov = "^4.0"
freezegun = "^1.2.2"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.idea
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
